+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * *[Beta6.rc1]*
   * [FIX]: The navigation is no longer disable-able on a device that
     does not support inline video
+  * [FEATURE]: Added VPAID and VAST companion ad support
 
 ## Beta5 (May 14, 2014)
 * *[Beta5.rc1]*